Changed the callback for a failed API call
The Retrofit Callback implemented in the MoviesListViewModel when
enqueuing the API call now simply logs the error and, additionally,
calls a UiErrorCallback, a SAM interface that should be implemented by
the View (Activity). The purpose of UiErrorCallback is to isolate in the
View layer the implementation of error notifications (e.g. a Toast
message or Snackbar).

Other changes:
 - Removed the MovieAdapter constructor. The class will fall back to the
  default constructor and then check for a null movies list before
  binding any ViewHolders.