Turned BodyLoggingInterceptor into a class
The BodyLoggingInterceptor was implemented as an anonymous class created
in a method inside the ResponseLoggingMovieDb class. It is more
appropriate for this Interceptor to be a class of its own, there is no
point in having a named anonymous class, especially if there's no reason
for it to be private to the class that uses it.

Also, the previous name of the class was ambigous (since a request can
also have a body), so it was renamed ResponseBodyLoggingInterceptor.

Other changes:
 - Removed the null arguments passed to the varargs methods in
 ResponseLoggingMovieDb.