Wrote class for encapsulating response handling
The ApiResponseHandler was created for encapsulating part of the
API-response-handling logic, as well as making the call site more
readable.

Readability is improved by requesting SAM interfaces for success and
error actions, thus enabling the use of lambda expressions at the call
site. Success here is defined as a successful API response i.e. code 200
with a body, and Error as any error (an unsuccessful or null-body
response is also considered an error).

The motivation for improving readability was the fact that the
Movie Detail ViewModel would need another retrofit2.Callback
implementation, and that it would be relatively similar to the one in
the Movies List ViewModel.