Changed Detail Activity to fill all its views
The MovieDetailActivity was changed to fill all its details views,
making use of all the info available in MovieDetails.

Some methods were created in the models (namely, MovieDetails and
Credits) to adapt info to the requirements of the Views e.g. provide a
List of Strings with the names of the directors, instead of a List of
directors as CrewMember instances.