Moved abbreviation logic to the ViewModel
The getLanguagesAbbreviated() and getCountriesAbbreviated() methods were
removed from the MovieDetails Model. Instead, part of this functionality
was implemented in the SpokenLanguage model, and another part in the
MovieDetailsViewModel. This two-part separation was architected
according to what concerns the Model, and what concerns the View Model
in terms of preparation of data for display. Thus the SpokenLanguage
Model merely contains Locale-related logic, but is not concerned with
any formatting. The View Model, on the other hand, contains formatting
logic, and is concerned, for instance, with deciding whether the
language name should be used instead of the abbreviation, according to
the number of languages given.

This architecture also takes care of the nonsense that was having the
MovieDetails Model return Lists of String, instead of Lists of
SpokenLanguage or ProductionCountry (Models), for the sole purpose of
enabling the use of the TextUtils methods in the View Model.
Additionally, TextUtils is no longer used at all, thus
minimizing the View Model's dependency of Android framework classes
(despite the fact that it is still an AndroidViewModel).

Other changes:
 - Some formatting details were adjusted in the MovieDetailsViewModel,
 such as appending "min" to the movie duration String returned to the
 View.
 - Some tools attributes were changed in the MovieDetails layout to
 reflect the formatting changes.