Set a transition between list and detail screens
A transition animation was set between the MoviesList and MovieDetails
Activities. To achieve this, some changes in the movie list item layout
were required so that the movie footer in the list returned to the
correct state when the transition was reversed. Additionally, the click
action setting that took place in the MovieAdapter has been reworked to
use an interface implemented by the Activity using it, having the
Adapter only call the listener passing the view and the movie ID.

The general implementation still needs changes, as 1, the animation is
actually linked to the whole detail layout, rather than the poster
only, deviating from the intended animation; and 2, the animation
doesn't wait for the details to be to be loaded.