Changed OnMovieClickListener
The OnMovieClickListener interface was specialized to use a
MoviePosterView parameter instead of a generic View. This expands
possibilities for click actions, since a special View will be expected.
Richer methods will be added to MoviePosterView, such as methods for
animating the info footer, which will make this interface change even
more useful. The interface was also renamed to reflect its single
method's signature change.