Fixed the info footer placement in movie posters
The info footer was corrected to display at the foot of the
MoviePosterView. The info footer would be placed at the top of the movie
poster, because it was being placed according to the parent layout
instead of the actual poster image ImageView. The parent layout has 0
height until Picasso loads the image, thus when loading finished the
footer remained at its original placement (y = 0). Now it updates
correctly when the ImageView's height becomes greater than 0.