Fixed error behavior in master-detail transitions
Previously, if a movie from the list was clicked and failed to load its
details in MovieDetailsActivity, the app would be apparently frozen (not
really) on this Activity, because the transition would only be triggered
when the details-loading process finished. Additionally, if the user
pressed the back button to return to the previous Activity, the
PosterView of the clicked movie would show up with an invisible footer
view.

This behavior was fixed by having the detail Activity finish when an
error occurs, and also by having the master Activity check on all its
PosterViews when it resumes. This way, each Activity is responsible for
its own integrity, thus keeping them decoupled on this front (sending
explicit intents between the two still couples them to some extent).