Changed 'net' package to accomodate local cache
The 'net' package war renamed and some of its classes were reorganized
and renamed, in order to prepare for the local cache functionality that
will be implemented. The communication point with the UI will be defined
by the interfaces in the 'data' package, namely Repository<T> and
DataRequest<T>, and networking and Retrofit implementation will be
abstracted from the other parts of the app.

Other changes:
 - Small code style rule change.