Implemented local db for offline caching
A local database for offline caching was implemented using Room.
Database design is far from ideal, using serialization to insert
complex object list columns, but this should work as an MVP, especially
because there won't be any major consequences when the schemas need to
be changed, as this db is for caching-only and the app isn't even in
production anyway.

The remoteapi -> localcache -> ui data flow takes place with a
LiveData<> object. First, an API call is enqueued with Retrofit for the
remote request, which will simply insert the data into the cache
database upon success. Next, a Room observable query LiveData<> is
immediately returned and passed to the UI. This LiveData<> will emit
with the first SELECT query return and with the INSERT resulting from
the remote API request, whichever happens first (or at all).

Other changes:
 - Changed wrapping settings for project style definition.