Fixed a crash caused by a method reference
Fixed a "Couldn't create instance of class AppMovieRepository" fatal
exception that was being thrown due to a method reference in one of the
SuccessCallback fields. For some reason this method couldn't be
reference, maybe because the 'cache' field hadn't been initialized yet.
A lambda expression doesn't cause this issue.

Other changes:
 - Made the AppMovieRepository constructor private to prevent external
 instantiation, and made the MoviesListViewModel an AndroidViewModel
 again, so that the application context can be used to get an instance
 of the Repository.